% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{Generic Plot function}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{object of class Rttest}

\item{...}{extra objects can be sent to method}
}
\value{
Returns a boxplot for two values using \code{ggplot()}
}
\description{
This function can be used to plot box plots for paired and  unpaired samples
}
\examples{
t = myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10)); plot(t)
}
