% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{Generic print function}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{object of class Rttest}

\item{...}{extra objects can be sent to method}
}
\value{
Prints list containing the confidence intervel and type of test
}
\description{
This function can be used to print the confidence interval for difference in means of two samples with it's confidence level and also prints the type of test being performed
}
\examples{
t = myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10)); print(t)
}
