% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyttest.R
\name{shinyttest}
\alias{shinyttest}
\title{Shiny Application}
\usage{
shinyttest()
}
\value{
The web page will display input sample data, summary of T-Test and box plot's of paired and unpaired samples
}
\description{
This function can be called to start shiny application and run ADVTTEST in reactive manner
}
\examples{
\dontrun{ shinyttest()}
}
