% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_t_test.r
\name{summary_t_test}
\alias{summary_t_test}
\title{Run a \emph{t} test for comparing means using summary statistics}
\usage{
summary_t_test(x1, x2, s1, s2, n1, n2, alternative = c("two.sided", "less",
  "greater"), method = c("pooled", "satterthwaite", "cochran"),
  conf.level = 0.95)
}
\arguments{
\item{x1, x2}{Sample means}

\item{s1, s2}{Sample standard devaitions}

\item{n1, n2}{Sample group sizes}

\item{alternative}{Upper, lower, or two-sided hypothesis}

\item{method}{Calculation method for degrees of freedom}

\item{conf.level}{Confidence level}
}
\value{
An object of class \code{"htest"}.
}
\description{
This gives you the ability to run \emph{t} tests to compare means using just
summary statistics, such as mean, standard deviation, and n. This is helpful
when the raw data underlying the summary statistics are not available.
}
\examples{
# Just plug in some summary data
summary_t_test(25,20,2,3,15,10, a = "l", m = "s")

# Using chickwts data from datasets
summary_chicks <- tapply(
  chickwts$weight,
  chickwts$feed,
  function(x) c(mean = mean(x), sd = sd(x), n = length(x))
)

# We now have a named list of vectors containing summary stats
summary_chicks

# Execute the t test using summary stats
with(summary_chicks,
     summary_t_test(
       linseed["mean"],
       soybean["mean"],
       linseed["sd"],
       soybean["sd"],
       linseed["n"],
       soybean["n"],
       alternative = "less",
       method = "sat"
       )
     )

# Compare with stats::t.test using raw data
t.test(
  weight ~ feed,
  chickwts,
  feed \%in\% c("linseed", "soybean"),
  alternative = "less"
)

}
\references{
P.V. Rao. (2007). \emph{Statistical Research Methods in the Life Sciences}.
pp 139-140. ISBN-13: 978-0-495-41422-3. ISBN-10: 0-495-41422-0
}
\author{
Paul W. Egeler, M.S., GStat
}
