% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{This is constructor function mttest}
\usage{
myttest(x, y, paired = FALSE, alpha = 0.05)
}
\arguments{
\item{x}{vector of the data}

\item{y}{vector of the data}

\item{paired}{Logical vector}

\item{alpha}{Significance level}
}
\value{
List of data, type of t-test and it's statistics
}
\description{
This is constructor function mttest
}
\examples{
myttest(x=rnorm(30,10,12), y=rnorm(40, 7, 10))
}
