% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyttest.R
\name{shinyttest}
\alias{shinyttest}
\title{Shiny Application}
\usage{
shinyttest()
}
\value{
Plots
}
\description{
This function can be called to start shiny application
}
\examples{
\dontrun{ shinyttest()}
}
